Folders reorganizeds and path aliases added and fixed
