+++ frontend/package.json
@@ -17,10 +17,12 @@
     "react-hook-form": "^7.62.0",
     "react-router": "^7.8.1",
     "react-spinners": "^0.17.0",
-    "tailwindcss": "^4.1.12"
+    "tailwindcss": "^4.1.12",
+    "vite-tsconfig-paths": "^5.1.4"
   },
   "devDependencies": {
     "@eslint/js": "^9.33.0",
+    "@types/node": "^24.3.0",
     "@types/react": "^19.1.10",
     "@types/react-dom": "^19.1.7",
     "@vitejs/plugin-react": "^5.0.0",
