+++ frontend/src/features/auth/components/RegisterForm.tsx
@@ -1,12 +1,12 @@
 import { useForm } from "react-hook-form";
-import { Input } from "../../../components/Input";
-import { Warning } from "../../../components/Warning";
-import Button from "../../../components/Button";
+import { Input } from "@/global/components/Input";
+import { Warning } from "@/global/components/Warning";
+import Button from "@/global/components/Button";
 import { useRegister } from "../hooks/useAuth";
 
 import { type RegisterFormProps } from "../types/RegisterFormType";
 import { type RegisterFormData } from "../types/RegisterFormType";
-import FormHeader from "../../../components/FormHeader";
+import FormHeader from "@/global/components/FormHeader";
 
 export default function RegisterForm({ className }: RegisterFormProps) {
   const { register, handleSubmit, formState, getValues } =
