+++ frontend/src/features/myAccount/components/DeleteAccountCard.tsx
@@ -1,6 +1,6 @@
 import { useState } from "react";
-import Button from "../../../components/Button";
-import FormHeader from "../../../components/FormHeader";
+import Button from "@/global/components/Button";
+import FormHeader from "@/global/components/FormHeader";
 import ConfirmationModal from "./ConfirmationModal";
 
 import { useDeleteAccount } from "../hooks/useDeleteAccount";
