+++ frontend/src/global/components/InputOnlyDisplayInfo.tsx
