+++ frontend/src/global/types/inputType.ts
