+++ frontend/vite.config.ts
@@ -2,13 +2,15 @@ import { defineConfig } from 'vite'
 import tailwindcss from '@tailwindcss/vite'
 import react from '@vitejs/plugin-react'
 import basicSsl from "@vitejs/plugin-basic-ssl"
+import tsconfigPaths from "vite-tsconfig-paths";
 
 // https://vite.dev/config/
 export default defineConfig({
   plugins: [
     tailwindcss(),
     react(),
-    basicSsl()
+    basicSsl(),
+    tsconfigPaths()
   ],
   server: {
     https: true,
